int -> long

+++ Mimir.Worker/StateDocumentConverter/AdventureCpStateDocumentConverter.cs
@@ -18,7 +18,7 @@ public class AdventureCpStateDocumentConverter : IStateDocumentConverter
                 $"{nameof(context.RawState)} Invalid state type. Expected {nameof(List)}, got {context.RawState.GetType().Name}."
             );
 
-        var cp = list[0].ToInteger();
+        var cp = list[0].ToLong();
 
         return new AdventureCpDocument(context.BlockIndex, context.Address, cp);
     }
